pyproject.toml
# ==================================================================================================
# Pyproject Configuration for the Loretex Package
# ==================================================================================================

# --- Build System Requirements --------------------------------------------------------------------

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"


# --- Package Metadata -----------------------------------------------------------------------------

[project]
name = "loretex"
version = "0.1.0"
authors = [{name = "Esther Poniatowski", email = "esther.poniatowski@ens.psl.eu"}]
license = "GPL-3.0-or-later"
license-files = ["LICENSE"]
readme = {file = "README.md", content-type = "text/markdown"}
description = "Converter between Markdown notes and LaTeX files"
keywords = ["markdown", "latex", "converter"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha"
]
urls = {homepage = "https://github.com/esther-poniatowski/loretex"}


# --- Dependencies ---------------------------------------------------------------------------------

requires-python = ">=3.12"

dependencies = [
    "pyyaml",            # YAML configuration parsing
    "attrs",             # Data classes and validation
    "typer",             # CLI interface
    "rich",              # Improved CLI output formatting
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "mypy>=1.8",
    "ruff>=0.1",
    "types-PyYAML",
]


# --- Package Source Code --------------------------------------------------------------------------

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
loretex = ["latex/*.sty", "latex/icons/pdf/*.pdf"]

# --- CLI Entry Points -----------------------------------------------------------------------------

[project.scripts]
loretex = "loretex.cli:app"
